Refactor
Made QuestionRoundIssuedService into a CohEventAction to match all the
other ways we handle CohEvents.
