+++ src/functionalTest/java/uk/gov/hmcts/reform/sscscorbackend/CohRequests.java
@@ -20,6 +20,7 @@ import uk.gov.hmcts.reform.sscs.idam.IdamTokens;
 import uk.gov.hmcts.reform.sscscorbackend.thirdparty.coh.api.CreateOnlineHearingRequest;
 
 public class CohRequests {
+    private static final long SECS_TO_WAIT_COH = 60L;
     private final IdamTokens idamTokens;
     private String cohBaseUrl;
     private HttpClient cohClient;
@@ -62,7 +63,7 @@ public class CohRequests {
                 "{\"state_name\": \"question_issue_pending\"}"
         );
 
-        waitUntil(roundIssued(hearingId), 60L, "Question round has not been issued in 60 seconds. Hearing id [" + hearingId + "]");
+        waitUntil(roundIssued(hearingId), SECS_TO_WAIT_COH, "Question round has not been issued in " + SECS_TO_WAIT_COH + " seconds. Hearing id [" + hearingId + "]");
     }
 
     public String createAnswer(String hearingId, String questionId, String answerText) throws IOException {
@@ -101,7 +102,7 @@ public class CohRequests {
                         "}"
         );
 
-        waitUntil(decisionIssued(hearingId), 60L, "Decision has not been issued in 60 seconds. Hearing id [" + hearingId + "]");
+        waitUntil(decisionIssued(hearingId), SECS_TO_WAIT_COH, "Decision has not been issued in " + SECS_TO_WAIT_COH + " seconds. Hearing id [" + hearingId + "]");
     }
 
     public int getDeadlineExtensionCount(String hearingId) throws IOException {
