+++ src/main/java/uk/gov/hmcts/reform/sscscorbackend/controllers/coheventmapper/QuestionRoundIssuedEventAction.java
@@ -1,33 +1,37 @@
-package uk.gov.hmcts.reform.sscscorbackend.service;
+package uk.gov.hmcts.reform.sscscorbackend.controllers.coheventmapper;
 
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.sscs.ccd.domain.SscsCaseDetails;
+import uk.gov.hmcts.reform.sscscorbackend.service.CorEmailService;
+import uk.gov.hmcts.reform.sscscorbackend.service.DwpEmailMessageBuilder;
+import uk.gov.hmcts.reform.sscscorbackend.service.StoreQuestionsPdfService;
 import uk.gov.hmcts.reform.sscscorbackend.service.pdf.StorePdfResult;
 import uk.gov.hmcts.reform.sscscorbackend.thirdparty.coh.apinotifications.CohEvent;
 import uk.gov.hmcts.reform.sscscorbackend.thirdparty.notifications.NotificationsService;
 
 @Service
-public class QuestionRoundIssuedService {
+public class QuestionRoundIssuedEventAction implements CohEventAction {
     private final NotificationsService notificationsService;
     private final StoreQuestionsPdfService storeQuestionsPdfService;
     private final CorEmailService corEmailService;
     private final DwpEmailMessageBuilder dwpEmailMessageBuilder;
 
-    public QuestionRoundIssuedService(NotificationsService notificationsService,
-                                      StoreQuestionsPdfService storeQuestionsPdfService,
-                                      CorEmailService corEmailService,
-                                      DwpEmailMessageBuilder dwpEmailMessageBuilder) {
+    public QuestionRoundIssuedEventAction(NotificationsService notificationsService,
+                                          StoreQuestionsPdfService storeQuestionsPdfService,
+                                          CorEmailService corEmailService,
+                                          DwpEmailMessageBuilder dwpEmailMessageBuilder) {
         this.notificationsService = notificationsService;
         this.storeQuestionsPdfService = storeQuestionsPdfService;
         this.corEmailService = corEmailService;
         this.dwpEmailMessageBuilder = dwpEmailMessageBuilder;
     }
 
-    public void handleQuestionRoundIssued(CohEvent cohEvent) {
+    @Override
+    public void handle(Long caseId, String onlineHearingId, CohEvent cohEvent) {
         notificationsService.send(cohEvent);
         StorePdfResult storePdfResult = storeQuestionsPdfService.storePdf(
-                Long.valueOf(cohEvent.getCaseId()),
-                cohEvent.getOnlineHearingId()
+                caseId,
+                onlineHearingId
         );
         SscsCaseDetails sscsCaseDetails = storePdfResult.getDocument();
         String caseReference = sscsCaseDetails.getData().getCaseReference();
